feat(tier4_planning_rviz_plugin): move footprint plugin to path (#2971)
* feat(tier4_rviz_plugin): simplify tier4_planning_rviz_plugin


* update


---------
