+++ common/tier4_planning_rviz_plugin/include/path/display.hpp
@@ -21,6 +21,8 @@
 #include <autoware_auto_planning_msgs/msg/path_with_lane_id.hpp>
 #include <autoware_auto_planning_msgs/msg/trajectory.hpp>
 
+#include <memory>
+#include <utility>
 #include <vector>
 
 namespace
@@ -158,6 +160,23 @@ class AutowarePathWithLaneIdDisplay
 : public AutowarePathWithDrivableAreaDisplay<autoware_auto_planning_msgs::msg::PathWithLaneId>
 {
   Q_OBJECT
+public:
+  AutowarePathWithLaneIdDisplay();
+  ~AutowarePathWithLaneIdDisplay();
+
+private:
+  void preprocessMessageDetail(
+    const autoware_auto_planning_msgs::msg::PathWithLaneId::ConstSharedPtr msg_ptr) override;
+  void processMessageDetail(
+    const autoware_auto_planning_msgs::msg::PathWithLaneId::ConstSharedPtr msg_ptr,
+    const size_t p_idx) override;
+
+  rviz_common::properties::BoolProperty property_lane_id_view_;
+  rviz_common::properties::FloatProperty property_lane_id_scale_;
+
+  using LaneIdObject =
+    std::pair<std::unique_ptr<Ogre::SceneNode>, std::unique_ptr<rviz_rendering::MovableText>>;
+  std::vector<LaneIdObject> lane_id_obj_ptrs_;
 };
 
 class AutowarePathDisplay
